# kardo-swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "KARDO API"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/admin/selections/offline/{selectionId}": {
      "delete": {
        "tags": [
          "Оффлайн-отбор: Admin."
        ],
        "summary": "Удаление оффлайн-отбора.",
        "operationId": "deleteOfflineSelection",
        "parameters": [
          {
            "name": "selectionId",
            "in": "path",
            "description": "id оффлайн-отбора",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Оффлайн-отбор не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Оффлайн-отбор удалён."
          }
        }
      },
      "patch": {
        "tags": [
          "Оффлайн-отбор: Admin."
        ],
        "summary": "Обновление оффлайн-отбора.",
        "operationId": "updateOfflineSelectionById",
        "parameters": [
          {
            "name": "selectionId",
            "in": "path",
            "description": "id оффлайн-отбора",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOfflineSelectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Оффлайн-отбор не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Оффлайн-отбор обновлен.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfflineSelectionDto"
                }
              }
            }
          }
        }
      }
    },
    "/admin/participations/selections/offline/{selectionId}": {
      "get": {
        "tags": [
          "Заявка на участие в отборе: Admin."
        ],
        "summary": "Получение списка заявок на участие в конкретном отборе.",
        "operationId": "getParticipationsBySelectionId",
        "parameters": [
          {
            "name": "selectionId",
            "in": "path",
            "description": "id Гранд-соревнования",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Количество элементов, которые нужно пропустить для формирования текущего набора",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Количество элементов в наборе",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Гранд-соревнование не найдено",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Список заявок на участие получен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ParticipationRequestDto"
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Заявка на участие в отборе: Admin."
        ],
        "summary": "Обновление заявок на участие в конкретном отборе.",
        "operationId": "updateParticipationRequestStatus",
        "parameters": [
          {
            "name": "selectionId",
            "in": "path",
            "description": "id отбора",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParticipationRequestStatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Отбор не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Заявки обновлены.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipationRequestStatusUpdateResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/selections/offline/{selectionId}": {
      "get": {
        "tags": [
          "Оффлайн-отбор: Users."
        ],
        "summary": "Получение оффлайн-отбора.",
        "operationId": "getOfflineSelectionById",
        "parameters": [
          {
            "name": "selectionId",
            "in": "path",
            "description": "id оффлайн-отбора",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Оффлайн-отбор не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Оффлайн-отбор найден.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfflineSelectionDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserShortDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "birthday": {
            "type": "string",
            "format": "date"
          },
          "country": {
            "type": "string"
          },
          "avatarPhoto": {
            "type": "string"
          }
        },
        "description": "Сущность пользователя короткая."
      },
...
      "GrandCompetitionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "hashtag": {
            "type": "string"
          },
          "competitionStart": {
            "type": "string",
            "format": "date"
          },
          "competitionEnd": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "FINISHED",
              "UPCOMING",
              "IN_PROGRESS"
            ]
          },
          "location": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "BREAKING",
                "PARKOUR",
                "WORKOUT",
                "KICKSCOOTERING",
                "TRICKING",
                "SKATEBOARDING",
                "BMX",
                "HIP_HOP",
                "DJ_ING",
                "GRAFFITI",
                "WAKEBOARDING",
                "ALL_FIELD"
              ]
            }
          },
          "description": {
            "type": "string"
          }
        },
        "description": "Сущность Гранд-соревнования."
      },
      "OfflineSelectionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "competitionDto": {
            "$ref": "#/components/schemas/GrandCompetitionDto"
          },
          "selectionStart": {
            "type": "string",
            "format": "date"
          },
          "selectionEnd": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "FINISHED",
              "UPCOMING",
              "IN_PROGRESS"
            ]
          },
          "selectionType": {
            "type": "string",
            "enum": [
              "REGIONAL",
              "NATIONAL"
            ]
          },
          "fields": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "BREAKING",
                "PARKOUR",
                "WORKOUT",
                "KICKSCOOTERING",
                "TRICKING",
                "SKATEBOARDING",
                "BMX",
                "HIP_HOP",
                "DJ_ING",
                "GRAFFITI",
                "WAKEBOARDING",
                "ALL_FIELD"
              ]
            }
          },
          "location": {
            "type": "string"
          }
        },
        "description": "Сущность оффлайн-отбора."
      },
...
      "ParticipationRequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "creationTime": {
            "type": "string",
            "format": "date-time"
          },
          "selectionDto": {
            "$ref": "#/components/schemas/OfflineSelectionDto"
          },
          "requesterDto": {
            "$ref": "#/components/schemas/UserShortDto"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "CANCELED"
            ]
          },
          "fields": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "BREAKING",
                "PARKOUR",
                "WORKOUT",
                "KICKSCOOTERING",
                "TRICKING",
                "SKATEBOARDING",
                "BMX",
                "HIP_HOP",
                "DJ_ING",
                "GRAFFITI",
                "WAKEBOARDING",
                "ALL_FIELD"
              ]
            }
          }
        },
        "description": "Сущность заявки на участие в отборе."
      },
...
      "UpdateOfflineSelectionRequest": {
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 250,
            "minLength": 2,
            "type": "string"
          },
          "selectionStart": {
            "type": "string",
            "format": "date"
          },
          "selectionEnd": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "FINISHED",
              "UPCOMING",
              "IN_PROGRESS"
            ]
          },
          "selectionType": {
            "type": "string",
            "enum": [
              "REGIONAL",
              "NATIONAL"
            ]
          },
          "fields": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "BREAKING",
                "PARKOUR",
                "WORKOUT",
                "KICKSCOOTERING",
                "TRICKING",
                "SKATEBOARDING",
                "BMX",
                "HIP_HOP",
                "DJ_ING",
                "GRAFFITI",
                "WAKEBOARDING",
                "ALL_FIELD"
              ]
            }
          },
          "location": {
            "maxLength": 250,
            "minLength": 2,
            "type": "string"
          }
        },
        "description": "Данные обновляемого отбора"
      },
      "ParticipationRequestStatusUpdateRequest": {
        "required": [
          "requestIds",
          "status"
        ],
        "type": "object",
        "properties": {
          "requestIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "CONFIRMED",
              "REJECTED"
            ]
          }
        },
        "description": "Данные для обновления заявок"
      },
      "ParticipationRequestStatusUpdateResult": {
        "type": "object",
        "properties": {
          "updatedRequests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipationRequestDto"
            }
          },
          "notUpdatedRequests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipationRequestDto"
            }
          }
        },
        "description": "Возвращаемая сущность обновления статуса заявки участника."
      },
...
    }
  }
}